+++ src/test/polymer/iron-flex-layout-classes_test.ts
@@ -85,7 +85,7 @@ Import them in the template style include.`,
   iron-flex-reverse iron-flex-factors
 
 Import them in the template style include.`,
-  'Style outside template. Run `move-style-into-template` rule.',
+      'Style outside template. Run `move-style-into-template` rule.',
       `Style module is used but not imported:
 
   iron-positioning
