+++ test/iron-flex-layout-classes/after-fixes.html
@@ -112,7 +112,27 @@
     </template>
     <script>
       Polymer({
-        is: "with-style"
+        is: "with-style-outside"
+      });
+    </script>
+  </dom-module>
+
+
+  <dom-module id="with-right-include">
+    <template>
+      <style include="iron-flex">
+         :host {
+          display: block;
+        }
+      </style>
+      <div class="flex">
+        <p>lorem</p>
+        <p>ipsum</p>
+      </div>
+    </template>
+    <script>
+      Polymer({
+        is: "with-right-include"
       });
     </script>
   </dom-module>
